feat(front): ajout du système d'upload img système simple coté frontend avec 2 call API

+++ src/app/formulaires/create-article/create-article.component.ts
@@ -7,6 +7,7 @@ import { ArticleService } from 'src/app/services/article.service';
 import { CategorieService } from 'src/app/services/categorie.service';
 import { DateTime } from 'luxon';
 import { Router } from '@angular/router';
+import { HttpClient } from '@angular/common/http';
 
 @Component({
   selector: 'app-create-article',
@@ -23,8 +24,11 @@ export class CreateArticleComponent implements OnInit{
   //je récupère la liste des catégories
   categories$: Observable<Categorie[]> =  this.categorieService.getCategories();
 
+  fileName: string = '';
+  file!: File;
+  uuid!: string;
+
   // boolean pour affichage de la validation de la requete
-  fichierValide: boolean = false;
   articleValide: boolean = false;
   articleDeleted: boolean = false;
 
@@ -36,8 +40,7 @@ export class CreateArticleComponent implements OnInit{
     categories: [[]],
     titre: ['', Validators.required], // je peux mettre un ou plusieurs validateur(s)
     corps: ['', Validators.required],
-    uuid: [''],
-    fichier: []   
+    fichier: [null]   
   });
 
   // formValues pour la suppression de la salle
@@ -57,7 +60,8 @@ export class CreateArticleComponent implements OnInit{
     private formBuilder: FormBuilder,
     private articleService: ArticleService,
     private categorieService: CategorieService,
-    private router:Router
+    private router:Router,
+    private http: HttpClient
     ){}
 
   ngOnInit(): void {
@@ -71,53 +75,68 @@ export class CreateArticleComponent implements OnInit{
     })
   }
 
+  onFileSelected(event: any) {
+    this.file = event.target.files[0];
+    if (this.file) {
+      this.fileName = this.file.name;
+      this.articleService.sendFile(this.file).subscribe({
+        next:(response:any) => {
+          console.log(response, !!response)
+          this.uuid=response.fichier;
+          console.log("onfileselected : "+ this.uuid);
+          
+        },
+        error:(error:any) => {
+          //throw erreur
+          console.log(error);
+        }
+      })
+    }
+  }
+
   /**
    * envoie les éléments de ll'article vers le service
    * @param e event du template
    */
   onAddArticle(formGroup: FormGroup) {
-    // debug
-    console.log(JSON.stringify(formGroup.value, null, 2));
 
     // je passe la variable submitted à true pour pouvoir afficher la confirmation à l'écran avec un ngIf
-    this.submitted = true;
+    this.submitted = true; 
+    const result = {
+      titre: formGroup.value.titre,
+      corps: formGroup.value.corps,
+      fichier: this.uuid,
+      categories: [{"id": formGroup.value.categories}],
+      like_dislike: 0
+    }
+    console.log(result);
+  
+    // // je mets la date du jour au bon format
+    // formGroup.value.date_ecriture = DateTime.now().toFormat('yyyy-MM-dd');
 
-    //récupération de uuid du ficher
-    this.articleService.sendFile(formGroup.value.fichier).subscribe({
-      next:(reponse:any) => {
-        this.fichierValide = true;
-        formGroup.value.uuid = reponse;
-      },
-      error:(error:any) => {
-        //throw erreur
-        console.log(error);
-      }
-    })
-    this.onCompletForm(formGroup);
-  }
-   
-  /**
-   * méthode qui complète le formulaire et la création de l'article après avoir
-   * récupéré l'uuid du fichier
-   * @param formGroup formulaire de l'article
-   */
-  onCompletForm(formGroup: FormGroup) {
-    // je mets la date du jour au bon format
-    formGroup.value.date_ecriture = DateTime.now().toFormat('yyyy-MM-dd');
+    // // je mets la catégorie dans l'objet à poster et complète l'objet article
+    // formGroup.value.categories = [{"id": formGroup.value.categories}];
+    // formGroup.value.date_modif = formGroup.value.date_ecriture;
+    // formGroup.value.like_dislike = 0
 
-    // je mets la catégorie dans l'objet à poster et complète l'objet article
-    formGroup.value.categories = [{"id": formGroup.value.categories}];
-    formGroup.value.date_modif = formGroup.value.date_ecriture;
-    formGroup.value.like_dislike = 0
+    // //ajout du fichier
+    // console.log("onaddarticle : "+this.uuid);
+    // formGroup.value.uuid = this.uuid;
+    // console.log("onaddarticle : "+this.uuid);
+
+    // formGroup.value.fichier = this.file;
 
     //  je vérifie si le formulaire est valide
     if (formGroup.valid) {
+      console.log("formgroup valid");
+      
       // si le formulaire est valide, je passe la variable formValidated à true ce qui me permettra de signaler
       // à l'utilisateur que le formulaire a bien été validé via un message
-      this.articleService.createArticle(formGroup.value).subscribe({
+      this.articleService.createArticle(result).subscribe({
         next:(response:any) => {
           this.articleValide=true;
-          window.location.reload();
+          this.articleList$ = this.articleService.getArticles();
+          
         },
         error:(error:any) => {
           //throw erreur
@@ -131,15 +150,13 @@ export class CreateArticleComponent implements OnInit{
    * suppression d'un article
    * @param id de l'article
    */
-  onDeleteArticle(id: number) {
+  onDeleteArticle(id?: number) {
     
     // je passe la variable submitted à true
     this.deleteSubmitted = true;
-      
-    this.articleService.deleteArticle(id).subscribe({
+    if (id) this.articleService.deleteArticle(id).subscribe({
       next:(response:any) => {
         this.articleDeleted=true;
-        window.location.reload();
       },
       error:(error:any) => {
         //throw erreur
@@ -152,7 +169,7 @@ export class CreateArticleComponent implements OnInit{
   alertFormValues(formGroup: FormGroup) {
     alert(JSON.stringify(formGroup.value, null, 2));
   }
-  onRedirectEditArticle(id: number){
+  onRedirectEditArticle(id?: number){
     this.router.navigate(['/modifierarticle', id])
   }
